Add storefront context directly to ACDL
